Match-id-fe0c24173b93e20e11d4d9220996bda1d1fe0971

+++ analysis/csrc/viewer/database/drafts/cann_trace_db_dumper.h
@@ -36,8 +36,8 @@ using HostTask = Analysis::Entities::HostTask;
 using OpDesc = Analysis::Entities::OpDesc;
 using HostTasks = std::vector<std::shared_ptr<HostTask>>;
 using TaskInfoData = std::vector<std::tuple<uint32_t, std::string, uint32_t, uint32_t, uint32_t, uint32_t, std::string,
-        std::string, std::string, int32_t, uint32_t, double, uint32_t, uint32_t, std::string, std::string, std::string,
-        std::string, std::string, std::string, uint32_t, uint32_t, std::string>>;
+        std::string, std::string, int32_t, uint32_t, uint64_t, uint32_t, uint32_t, std::string, std::string,
+        std::string, std::string, std::string, std::string, uint32_t, uint32_t, std::string>>;
 using HCCLBigOpDescs = Analysis::Association::Cann::HCCLBigOpDescs;
 public:
     // 创建时传入host路径
